# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/sales_orders:
    get:
      tags:
        - SalesOrders
      summary: List SalesOrders
      description: Get a paginated list of SalesOrders with optional filtering and sorting
      operationId: listSalesOrders
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesOrderList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - SalesOrders
      summary: Create SalesOrder
      description: Create a new SalesOrder
      operationId: createSalesOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSalesOrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/sales_order_items:
    get:
      tags:
        - SalesOrderItems
      summary: List SalesOrderItems
      description: Get a paginated list of SalesOrderItems with optional filtering and sorting
      operationId: listSalesOrderItems
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesOrderItemList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - SalesOrderItems
      summary: Create SalesOrderItem
      description: Create a new SalesOrderItem
      operationId: createSalesOrderItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSalesOrderItemRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesOrderItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    SalesOrder:
      type: object
      required:
        - id
        - order_number
        - company_id
        - customer_id
        - status
        - order_date
        - currency
        - subtotal
        - tax_rate
        - tax_amount
        - total
        - metadata
      properties:
        id:
          type: string
          format: uuid
        order_number:
          type: string
          maxLength: 40
        quotation_id:
          nullable: true
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        status:
          type: string
        order_date:
          type: string
          format: date
        delivery_date:
          nullable: true
          type: string
          format: date
        currency:
          type: string
          maxLength: 3
        subtotal:
          type: string
        tax_rate:
          type: string
        tax_amount:
          type: string
        total:
          type: string
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    CreateSalesOrderRequest:
      type: object
      required:
        - order_number
        - company_id
        - customer_id
        - status
        - order_date
        - currency
        - subtotal
        - tax_rate
        - tax_amount
        - total
        - metadata
      properties:
        order_number:
          type: string
          maxLength: 40
        quotation_id:
          nullable: true
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        status:
          type: string
        order_date:
          type: string
          format: date
        delivery_date:
          nullable: true
          type: string
          format: date
        currency:
          type: string
          maxLength: 3
        subtotal:
          type: string
        tax_rate:
          type: string
        tax_amount:
          type: string
        total:
          type: string
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    SalesOrderList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalesOrder'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    SalesOrderItem:
      type: object
      required:
        - id
        - order_id
        - company_id
        - item_id
        - quantity
        - unit_price
        - line_discount
        - line_amount
        - billed_qty
        - delivered_qty
        - metadata
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 300
        quantity:
          type: string
        unit_price:
          type: string
        line_discount:
          type: string
        line_amount:
          type: string
        billed_qty:
          type: string
        delivered_qty:
          type: string
        metadata:
          type: object
...
    CreateSalesOrderItemRequest:
      type: object
      required:
        - order_id
        - company_id
        - item_id
        - quantity
        - unit_price
        - line_discount
        - line_amount
        - billed_qty
        - delivered_qty
        - metadata
      properties:
        order_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 300
        quantity:
          type: string
        unit_price:
          type: string
        line_discount:
          type: string
        line_amount:
          type: string
        billed_qty:
          type: string
        delivered_qty:
          type: string
        metadata:
          type: object
...
    SalesOrderItemList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalesOrderItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []